# frontend/src/generated/openapi.v1.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Grocery Store SOP Assistant API",
    "description": "RAG-powered chatbot API for the Grocery Store Standard Operating Procedures.",
    "version": "1.0"
  },
  "paths": {
...
    "/api/v1/Chat/stream": {
      "post": {
        "tags": ["Chat"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/Chat/feedback": {
      "post": {
        "tags": ["Chat"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConversationFeedbackRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ConversationFeedbackRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ConversationFeedbackRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ConversationFeedbackResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConversationFeedbackResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConversationFeedbackResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ChatMessageDto": {
        "required": ["content", "role"],
        "type": "object",
        "properties": {
          "role": {
            "minLength": 1,
            "pattern": "^(user|assistant|system)$",
            "type": "string"
          },
          "content": {
            "maxLength": 4000,
            "minLength": 1,
            "type": "string"
          },
          "timestampUtc": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "ChatRequest": {
        "required": ["conversationId", "messages"],
        "type": "object",
        "properties": {
          "conversationId": {
            "maxLength": 128,
            "minLength": 1,
            "type": "string"
          },
          "messages": {
            "maxItems": 20,
            "minItems": 1,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ChatMessageDto"
            }
          },
          "knowledgeBaseId": {
            "maxLength": 64,
            "pattern": "^[a-zA-Z0-9._-]*$",
            "type": "string",
            "nullable": true
          },
          "userRole": {
            "maxLength": 32,
            "pattern": "^(cashier|manager|department_lead)?$",
            "type": "string",
            "nullable": true
          },
          "useTools": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
      "ConversationFeedbackRequest": {
        "required": ["conversationId", "feedbackType", "messageId"],
        "type": "object",
        "properties": {
          "conversationId": {
            "maxLength": 128,
            "minLength": 1,
            "type": "string"
          },
          "messageId": {
            "maxLength": 128,
            "minLength": 1,
            "type": "string"
          },
          "feedbackType": {
            "minLength": 1,
            "pattern": "^(helpful|unhelpful|wrong-citation)$",
            "type": "string"
          },
          "comment": {
            "maxLength": 500,
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ConversationFeedbackResponse": {
        "type": "object",
        "properties": {
          "accepted": {
            "type": "boolean"
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "submittedAtUtc": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
...
    }
  },
...
}